+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Add
 - Add `IMGPROXY_FALLBACK_IMAGE_TTL` config.
+- Add [watermark_size](https://docs.imgproxy.net/generating_the_url?id=watermark-size) processing option.
 
 ### Change
 - (pro) Don't check `Content-Length` header of videos.