+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 ### Add
 - Add `IMGPROXY_FALLBACK_IMAGE_TTL` config.
-- Add [watermark_size](https://docs.imgproxy.net/generating_the_url?id=watermark-size) processing option.
+- (pro) Add [watermark_size](https://docs.imgproxy.net/generating_the_url?id=watermark-size) processing option.
 - Add OpenStack Object Storage ("Swift") support.
 
 ### Change